requirements.txt
Flask==2.0.3
flask-cors==3.0.10
pymongo==4.0.2
python-dotenv==0.19.2
